+++ app/routers/morphology.py
@@ -1,11 +1,9 @@
-from typing import Annotated, NotRequired, TypedDict
+from typing import Annotated
 
 import sqlalchemy as sa
 from fastapi import APIRouter
 from fastapi_filter import FilterDepends
 from sqlalchemy.orm import (
-    InstrumentedAttribute,
-    Session,
     aliased,
     joinedload,
     raiseload,
@@ -26,23 +24,18 @@ from app.dependencies.common import PaginationQuery
 from app.dependencies.db import SessionDep
 from app.errors import ensure_result
 from app.filters.morphology import MorphologyFilter
+from app.routers.common import FacetQueryParams, get_facets
 from app.schemas.morphology import (
     ReconstructionMorphologyAnnotationExpandedRead,
     ReconstructionMorphologyCreate,
     ReconstructionMorphologyRead,
 )
-from app.schemas.types import Facet, Facets, ListResponse, PaginationResponse
+from app.schemas.types import ListResponse, PaginationResponse
 
 router = APIRouter(
     prefix="/reconstruction-morphology",
     tags=["reconstruction-morphology"],
 )
-
-
-class FacetQueryParams(TypedDict):
-    id: InstrumentedAttribute[int]
-    label: InstrumentedAttribute[str]
-    type: NotRequired[InstrumentedAttribute[str]]
 
 
 @router.get(
@@ -106,36 +99,6 @@ def create_reconstruction_morphology(
     return db_rm
 
 
-def _get_facets(
-    db: Session,
-    query: sa.Select,
-    name_to_facet_query_params: dict[str, FacetQueryParams],
-    count_distinct_field: InstrumentedAttribute,
-) -> Facets:
-    facets = {}
-    groupby_keys = ["id", "label", "type"]
-    orderby_keys = ["label"]
-    for facet_type, fields in name_to_facet_query_params.items():
-        groupby_fields = {"type": sa.literal(facet_type), **fields}
-        groupby_columns = [groupby_fields[key].label(key) for key in groupby_keys]  # type: ignore[attr-defined]
-        groupby_ids = [sa.literal(i + 1) for i in range(len(groupby_columns))]
-        facet_q = (
-            query.with_only_columns(
-                *groupby_columns,
-                sa.func.count(sa.func.distinct(count_distinct_field)).label("count"),
-            )
-            .group_by(*groupby_ids)
-            .order_by(*orderby_keys)
-        )
-        facets[facet_type] = [
-            Facet.model_validate(row, from_attributes=True)
-            for row in db.execute(facet_q).all()
-            if row.id is not None  # exclude null rows if present
-        ]
-
-    return facets
-
-
 @router.get("")
 def morphology_query(
     *,
@@ -180,7 +143,7 @@ def morphology_query(
     filter_query = morphology_filter.filter(filter_query, aliases={Agent: agent_alias})
 
     if with_facets:
-        facets = _get_facets(
+        facets = get_facets(
             db,
             filter_query,
             name_to_facet_query_params=name_to_facet_query_params,
