+++ tests/test_emodel.py
@@ -1,10 +1,6 @@
-import itertools as it
-
 import pytest
 
-from app.db.model import EModel, Species, Strain
-
-from .utils import BEARER_TOKEN, PROJECT_HEADERS, add_db, create_reconstruction_morphology_id
+from .utils import BEARER_TOKEN, PROJECT_HEADERS
 
 ROUTE = "/emodel"
 
@@ -47,7 +43,7 @@ def test_create_emodel(client, species_id, strain_id, brain_region_id, exemplar_
 
 
 # @pytest.mark.usefixtures("skip_project_check")
-# def test_query_reconstruction_morphology(db, client, brain_region_id):  # noqa: PLR0915
+# def test_query_reconstruction_morphology(db, client, brain_region_id):
 #     species1 = add_db(db, Species(name="TestSpecies1", taxonomy_id="0"))
 #     species2 = add_db(db, Species(name="TestSpecies2", taxonomy_id="1"))
